decoderConfigLen can be larger than 255 bytes
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@18667 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libmpdemux/parse_mp4.h
@@ -87,7 +87,7 @@ typedef struct {
   uint32_t avgBitrate;
 
   /* 0x05 DecSpecificDescrTag */
-  uint8_t  decoderConfigLen;
+  uint16_t  decoderConfigLen;
   uint8_t *decoderConfig;
 
   /* 0x06 SLConfigDescrTag */
